% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush_plots_binary.R
\name{brush_plots_binary}
\alias{brush_plots_binary}
\title{generate a shiny app for binary data type}
\usage{
brush_plots_binary(df, stats_df, group_list, group, value)
}
\arguments{
\item{df}{the processed data frame}

\item{stats_df}{the output table from fit_statistics function}

\item{group_list}{a list of distinct observations}

\item{group}{the column name for grouping variable}

\item{value}{the column name for response variable}
}
\value{
a shiny app
}
\description{
This is the method to generate a shiny app displaying the histograms of statistics and scatter plots and a table
of selected observations.
}
