% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush_plots_comparison.R
\name{brush_plots_comparison}
\alias{brush_plots_comparison}
\title{Interactive scatterplot to compare statistics}
\usage{
brush_plots_comparison(df, comparison_df, group_list, group, axes_vars, value)
}
\description{
Interactive scatterplot to compare statistics
}
