% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush_plots_comparison.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Faceted scatterplot of compared statistics}
\usage{
comparison_plot(comparison_df, group, axes_vars, cur_ids)
}
\description{
Faceted scatterplot of compared statistics
}
