% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_stats_histogram.R
\name{draw_stats_histogram}
\alias{draw_stats_histogram}
\title{draw histograms of statistics for each term}
\usage{
draw_stats_histogram(stats_df)
}
\arguments{
\item{stats_df}{the output table from fit_statistics function}
}
\value{
several histograms of all subsets based on terms in the model
}
\description{
This is a part of the shiny app to draw several histograms of statistics for each term
}
