% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit_statistics}
\alias{fit_statistics}
\title{fit_statistics}
\usage{
fit_statistics(subsets, lm_func, group)
}
\arguments{
\item{subsets}{the groups of subset from split_dataset function}

\item{lm_func}{the linear model chosen by user to fit the data}

\item{group}{the name of the group column}
}
\value{
a table with statistics of each term for all subsets
}
\description{
This method helps to fit a model for groups of subset
}
