% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush_plots_comparison.R
\name{reshape_comparisons}
\alias{reshape_comparisons}
\title{Reshape output from tests to enable comparison}
\usage{
reshape_comparisons(comparison_df, group, cur_ids, axes_vars, cur_var)
}
\description{
Reshape output from tests to enable comparison
}
