% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataset.R
\name{split_dataset}
\alias{split_dataset}
\title{Split the data set}
\usage{
split_dataset(df, group)
}
\arguments{
\item{df}{the given data frame}

\item{group}{the name of the group column}
}
\value{
groups of subsets
}
\description{
This function splits the given data set into many smaller subsets based on the group
}
